+++ src/Compute/Compute/ChangeLog.md
@@ -23,6 +23,7 @@
 ## Version 4.2.1
 * Added warning when using `New-AzVmss` without "latest" image version
 * Added '-Location' as optional positional parameter to Get-AzComputeResourceSku cmdlet
+* Added '-EncryptionAtHost' parameter to New-AzVm, New-AzVMConfig and Update-AzVM
 
 ## Version 4.2.0
 * Added SimulateEviction parameter to Set-AzVM and Set-AzVmssVM cmdlets.